refactor: move package post-processing into `project_support`
